move navigationMenu to containers folder
